% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{create_feature_mat}
\alias{create_feature_mat}
\title{Extract features mat}
\usage{
create_feature_mat(scmat, features = NULL,
  factor_by_median = get_param("scc_factor_by_median"),
  norm_size = get_param("scc_norm_size"),
  k_scale_umi = get_param("scc_k_scale_umi"), norm_features = T)
}
\arguments{
\item{scmat}{A tgScMat expression matrix.}

\item{features}{Either  a matrix of cell features (gene expression, PCA principle componant, etc.),
or a list of markers to be extracted from scmat. Clustering will be performed on these features.
If left blank, markers will be selected using the default method.}

\item{factor_by_median}{}

\item{norm_size}{}

\item{k_scale_umi}{}

\item{norm_features}{}
}
\value{
A matrix with one row per marker gene, one column per cell.
Gene expression is normalized according to the parameters.
}
\description{
Extract features mat
}
