% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_layout.r
\name{import_layout}
\alias{import_layout}
\title{Read a clustering layout from file.}
\usage{
import_layout(file, scl = NULL, ...)
}
\arguments{
\item{file}{Name of the file for inputting.}

\item{scl}{an instance of the relevant \code{\linkS4class{tgScMatClust}}, optional.}
}
\description{
Read a clustering layout from file.
}
