% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_layout.r
\name{plot_cells_breakdown_to_mods_by_feature}
\alias{plot_cells_breakdown_to_mods_by_feature}
\title{Plot heatmap of number of cells per cell module, breakdown by a metadata field, with additional metadata attributes}
\usage{
plot_cells_breakdown_to_mods_by_feature(sc_2d, outdir = get_param("outdir"),
  by_field = get_param("scp_cell_mod_tab_by"),
  metadata_fn = get_param("scm_batch_table_fn"),
  key_fields = get_param("scp_cell_mod_tab_key_fields"),
  add_text_to = get_param("scp_cell_mod_tab_add_txt_to"),
  key_dict = get_param("scp_cell_mod_tab_key_dict"),
  na_col = get_param("scp_cell_mod_tab_na_col"),
  cols_list = get_param("scp_cell_mod_tab_keys_col_dict"),
  show_legends = get_param("scp_cell_mod_tab_show_legend_for"),
  hm_pal = RColorBrewer::brewer.pal(n = 9, "YlOrRd"), n_shades = 200,
  cex_text = 1.5)
}
\arguments{
\item{sc_2d}{}

\item{outdir}{}

\item{metadata_fn}{}

\item{key_fields}{}

\item{add_text_to}{}

\item{key_dict}{}

\item{na_col}{}

\item{cols_list}{}

\item{show_legends}{}

\item{hm_pal}{}

\item{n_shades}{}

\item{cex_text}{}

\item{by}{}
}
\description{
Plot heatmap of number of cells per cell module, breakdown by a metadata field, with additional metadata attributes
}
