% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_select.r
\name{plot_markers}
\alias{plot_markers}
\title{plot statistics about the chosen markers}
\usage{
plot_markers(mat, marks, gstat, fname, outdir,
  shades = colorRampPalette(c("navy", "white", "brown"))(100),
  dendrogram.lab.cex = 2, h = 4500, w = 1800)
}
\description{
plot statistics about the chosen markers
}
