% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorter.r
\name{sc_marker_clusts}
\alias{sc_marker_clusts}
\title{cluster mat projected on markers genes}
\usage{
sc_marker_clusts(mat, markers_fn, factor_by_median = T, norm_size = T,
  k_scale_umi = 7, n_mark_clusts = 120, analysis_dir = getwd(),
  alg_type = "kmeans", min_clust_size = 20,
  tab_clust_fp_fn = "clust_fp.txt", clust_fp_metadata_fields = c("Patient",
  "Sample.Name", "Cell.type"),
  min_max_clust_fp_to_report = get_param("scc_min_max_clust_fp_to_report"),
  ordered_subtypes = c("ab_T_cell", "regulatory_CD4_T_cell", "CD4_T_cells",
  "CD8_T_cells", "gd_T_cell", "NK_cell", "cytotoxic_B_cells", "IgA_B_cell",
  "IgM_B_cell", "IgG_B_cell", "plasma_cells", "DC", "macrophage",
  "granulocytes", "erythrocyte", "osteoclast"), type_col = list(CD3 =
  "#e41a1c", CD45 = "#377eb8", `CD3&CD45` = "#4daf4a"), totmark_cl = NULL)
}
\arguments{
\item{mat}{}

\item{markers_fn}{}

\item{n_mark_clusts}{}

\item{analysis_dir}{}

\item{alg_type}{}

\item{min_clust_size}{}

\item{tab_clust_fp_fn}{}

\item{clust_fp_metadata_fields}{}
}
\value{

}
\description{
cluster mat projected on markers genes
}
