% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorter.r
\name{sc_marker_split}
\alias{sc_marker_split}
\title{assign cells to types by given projected markers cluster centers}
\usage{
sc_marker_split(mat, markers_fn, ref_centers_fn, ref_centers_type_assign_fn,
  factor_by_median = T, norm_size = T, k_scale_umi = 7, fig_pref = NULL,
  out_base_dir = NULL)
}
\arguments{
\item{mat}{}

\item{markers_fn}{gene to type table, project umi table onto types. If null, use the colnames of ref_centers_fn as markers and calc feature matrix}

\item{ref_centers_fn}{}

\item{ref_centers_type_assign_fn}{}

\item{factor_by_median}{}

\item{norm_size}{}

\item{k_scale_umi}{}

\item{fig_pref}{}

\item{out_base_dir}{}
}
\value{

}
\description{
assign cells to types by given projected markers cluster centers
}
