% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{sc_pipe_build_mat}
\alias{sc_pipe_build_mat}
\title{scrdb pipeline}
\usage{
sc_pipe_build_mat(scdb, index_fn = get_param("scm_batch_table_fn"),
  base_dir = get_param("scm_mars_base_dir"),
  batch_meta_attr = get_param("scm_batch_meta_attr"),
  mars_batches = get_param("scm_mars_batches"),
  sample_n_batches = get_param("scm_n_batches_to_sample"),
  outdir = get_param("outdir"), outdir_add_timestamp = F,
  min_umi_n = get_param("scm_min_cell_umis"),
  min_umis_init = min(get_param("scm_min_cell_umis_init"), min_umi_n),
  dataset_type = get_param("scm_dataset_type"),
  tenex_mat_ifn = get_param("scm_10x_mat_ifn"),
  tenex_batches = get_param("scm_10x_batches"))
}
\arguments{
\item{scdb}{initialized scdb object}

\item{index_fn}{}

\item{base_dir}{}

\item{batch_meta_attr}{}

\item{mars_batches}{}

\item{sample_n_batches}{}

\item{outdir}{}

\item{outdir_add_timestamp}{}

\item{min_umi_n}{}

\item{min_umis_init}{}
}
\value{
returns invisibly a tgScMat
}
\description{
pipeline for reading mars-seq, cleaning, finding cell and gene modules, 
and plotting results.
}
\details{
sc_pipe_build_mat

reading mars-seq into a single umi matrix
}
