% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{sc_pipe_cmods_to_gmods}
\alias{sc_pipe_cmods_to_gmods}
\title{Find correletaed gene modules on cell modules}
\usage{
sc_pipe_cmods_to_gmods(scdb, sc_cl, feat_type = "clust_fp",
  outdir = get_param("outdir"), n_mods = get_param("scgm_n_gmods"),
  foc_genes = get_param("scgm_foc_genes"),
  min_abs_log2_fold = get_param("scgm_min_abs_log2_fold"),
  min_tot = get_param("scgm_min_tot_umis"), rseed = 1)
}
\arguments{
\item{scdb}{}

\item{sc_cl}{}

\item{feat_type}{either clust_fp (correlate genes in clust_fp matrix) or residual (correlate genes in residual matrix after cleaning cluster specific expression)}

\item{outdir}{}

\item{foc_genes}{list of genes to use (null by default, and then finding genes by parameters below)}

\item{min_abs_log2_fold}{if feat_type is clust_fp, select genes with abs log2 fold change above this value}

\item{min_tot}{if feat_type is residual, select genes with total umi coutn above this value}

\item{rseed}{}
}
\value{
invisible tgScMatClust clustering object
}
\description{
Find correletaed gene modules on cell modules
}
