% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{sc_pipe_coclust_to_cell_modules}
\alias{sc_pipe_coclust_to_cell_modules}
\title{Generate new cell modules by cut tree on a cells' co-clust matrix (does not save the new cmods object in scdb)}
\usage{
sc_pipe_coclust_to_cell_modules(scdb, sc_cl,
  k_cut = get_param("scc_bootstrap_k_cut"),
  min_clust_size = get_param("scc_bootstrap_min_clust_size"),
  method = get_param("scc_bootstrap_clust_method"),
  outdir = get_param("outdir"),
  tab_clust_fp_fn = get_param("scc_clust_fp_fn"),
  min_max_clust_fp_to_report = get_param("scc_min_max_clust_fp_to_report"),
  clust_fp_metadata_fields = get_param("scc_clust_fp_metadata_fields"))
}
\arguments{
\item{scdb}{}

\item{sc_cl}{}

\item{k_cut}{}

\item{min_clust_size}{}
}
\value{
new tgScMatClust object
}
\description{
Generate new cell modules by cut tree on a cells' co-clust matrix (does not save the new cmods object in scdb)
}
