% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{sc_pipe_init_sort}
\alias{sc_pipe_init_sort}
\title{sc_pipe_init_sort}
\usage{
sc_pipe_init_sort(sc_mat, markers_fn,
  factor_by_median = get_param("scc_factor_by_median"),
  norm_size = get_param("scc_norm_size"),
  k_scale_umi = get_param("scc_k_scale_umi"), alg_type = "knn",
  min_clust_size = 20, n_mark_clusts = 80,
  ordered_subtypes = c("ab_T_cell", "regulatory_CD4_T_cell", "CD4_T_cells",
  "CD8_T_cells", "gd_T_cell", "NK_cell", "cytotoxic_B_cells", "IgA_B_cell",
  "IgM_B_cell", "IgG_B_cell", "plasma_cells", "DC", "macrophage",
  "granulocytes", "erythrocyte", "osteoclast"), sc_cl = NULL)
}
\arguments{
\item{sc_mat}{}

\item{markers_fn}{table with columns: subtype,gene,type}

\item{alg_type}{}

\item{min_clust_size}{}

\item{n_mark_clusts}{}

\item{ordered_subtypes}{}

\item{sc_cl}{}

\item{max_umi_count}{}
}
\value{
kmeans object of the projected mat
}
\description{
Project sc_mat on given cell type markers, cluster projected mat. Should be followed by manual
assignment of cell types to cluster centers
}
