% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{sc_pipe_mat_to_clean_mat}
\alias{sc_pipe_mat_to_clean_mat}
\title{sc_pipe_mat_to_clean_mat}
\usage{
sc_pipe_mat_to_clean_mat(scdb, raw_sc_mat, outdir = get_param("outdir"),
  outdir_add_timestamp = F,
  batch_meta_attr = get_param("scm_batch_meta_attr"),
  clust_knn = get_param("scc_k_knn"),
  min_umi_n = get_param("scm_min_cell_umis"),
  min_umis_init = min(get_param("scm_min_cell_umis_init"), min_umi_n),
  mark.min_var_mean = get_param("scc_marker_min_var_mean"),
  mark.sz_cor_norm_max = get_param("scc_marker_sz_cor_norm_max"),
  mark.niche_T = get_param("scc_marker_niche_T"),
  mark.min_tot = get_param("scc_marker_min_tot"),
  glob_blacklist_terms = get_param("scm_gene_blacklist_names"),
  mark_blacklist_terms = get_param("scc_marker_blacklist_names"),
  mark_blacklist_genes = get_param("scc_marker_blacklist_genes"),
  outlier_gene_top1 = get_param("scm_outlier_max_gene_top1"),
  outlier_gene_on_count = get_param("scm_outlier_gene_on_count"),
  remove_outliers_before_clustering = get_param("scm_remove_outliers"),
  min_clust_size = get_param("scc_min_clust_size"),
  filt_amb_on_clusts = get_param("scm_filter_amb_noise_on_clusts"),
  filt_outliers_on_clusts = get_param("scm_filter_outliers_on_clusts"),
  amb_epsilon = get_param("scm_amb_noise_epsilon"))
}
\arguments{
\item{scdb}{}

\item{raw_sc_mat}{}

\item{outdir}{}

\item{outdir_add_timestamp}{}

\item{batch_meta_attr}{}

\item{clust_knn}{}

\item{min_umi_n}{}

\item{min_umis_init}{}

\item{mark.min_var_mean}{}

\item{mark.sz_cor_norm_max}{}

\item{mark.niche_T}{}

\item{mark.min_tot}{}

\item{glob_blacklist_terms}{}

\item{mark_blacklist_terms}{regular expression terms to search blacklisted genes}

\item{mark_blacklist_genes}{exact names of blacklist genes}

\item{outlier_gene_top1}{}

\item{outlier_gene_on_count}{}

\item{remove_outliers_before_clustering}{}

\item{min_clust_size}{}

\item{filt_amb_on_clusts}{}

\item{filt_outliers_on_clusts}{}

\item{amb_epsilon}{}
}
\value{
returns invisibly a tgScMat
}
\description{
cleaning the umi table
}
