% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{sc_pipe_mat_to_cmods}
\alias{sc_pipe_mat_to_cmods}
\title{sc_pipe_mat_to_cmods}
\usage{
sc_pipe_mat_to_cmods(scdb, sc_mat, outdir = get_param("outdir"),
  outdir_add_timestamp = F, use_bootstapping = get_param("scc_bootstrap"),
  k_cut = get_param("scc_bootstrap_k_cut"),
  N_boot = get_param("scc_bootstrap_n_boot"),
  boot_ratio = get_param("scc_bootstrap_ratio"),
  n_procs = get_param("scc_bootstrap_n_procs"),
  boot_min_clust_size = get_param("scc_bootstrap_min_clust_size"),
  clust_knn = get_param("scc_k_knn"),
  mark.min_var_mean = get_param("scc_marker_min_var_mean"),
  mark.sz_cor_norm_max = get_param("scc_marker_sz_cor_norm_max"),
  mark.niche_T = get_param("scc_marker_niche_T"),
  mark.min_tot = get_param("scc_marker_min_tot"),
  mark_blacklist_terms = get_param("scc_marker_blacklist_names"),
  mark_blacklist_genes = get_param("scc_marker_blacklist_genes"),
  filt_outliers_on_clusts = get_param("scm_filter_outliers_on_clusts"),
  min_clust_size = get_param("scc_min_clust_size"),
  tab_clust_fp_fn = get_param("scc_clust_fp_fn"),
  min_max_clust_fp_to_report = get_param("scc_min_max_clust_fp_to_report"),
  clust_fp_metadata_fields = get_param("scc_clust_fp_metadata_fields"))
}
\arguments{
\item{scdb}{}

\item{sc_mat}{}

\item{outdir}{}

\item{outdir_add_timestamp}{}

\item{use_bootstapping}{}

\item{k_cut}{}

\item{N_boot}{}

\item{boot_ratio}{}

\item{clust_knn}{}

\item{mark.min_var_mean}{}

\item{mark.sz_cor_norm_max}{}

\item{mark.niche_T}{}

\item{mark.min_tot}{}

\item{mark_blacklist_genes}{}

\item{filt_outliers_on_clusts}{}

\item{min_clust_size}{}

\item{tab_clust_fp_fn}{}

\item{clust_fp_metadata_fields}{}

\item{n_proces}{}

\item{boot_in_clust_size}{}
}
\value{
invisible tgScMatClust clustering object
}
\description{
cluster input sc_mat
}
