% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{sc_pipe_plots}
\alias{sc_pipe_plots}
\title{sc_pipe_plots}
\usage{
sc_pipe_plots(scdb, sc_cl, outdir = get_param("outdir"),
  mark_blacklist_terms = get_param("scc_marker_blacklist_names"),
  mark_blacklist_genes = get_param("scc_marker_blacklist_genes"),
  mark_colors_fn = get_param("scp_mark_colors_fn"),
  fig_width_confu = get_param("scp_width_confu"),
  fig_mat_per_clust_genes = get_param("scp_ord_per_clust_genes"),
  fig_mat_gene_min_fold = get_param("scp_ord_gene_min_fold"),
  fig_mat_gene_min_cov = get_param("scp_ord_gene_min_cov"),
  fig_mat_width = get_param("scp_mat_width"),
  fig_mat_height = get_param("scp_mat_height"),
  fig_mat_text_cex = get_param("scp_mat_text_cex"),
  fig_mat_smooth_n = get_param("scp_mat_smooth_n"),
  fig_2d_height = get_param("scp_2d_height"),
  fig_2d_width = get_param("scp_2d_width"),
  K_2dproj = get_param("scp_K_2dproj"),
  K_cells_2dproj = get_param("scp_K_cells_2dproj"),
  force_max_deg_2dproj = get_param("scp_force_max_deg_2dproj"),
  T_edge_2dproj = get_param("scp_T_edge_2dproj"),
  T_edge_asym = get_param("scp_T_edge_asym"),
  restrict_edges_by_fpcor = get_param("scp_restrict_edges_by_fpcor"),
  expand_K_connect_2dproj = get_param("scp_expand_K_connect_2dproj"),
  fold_shades = colorRampPalette(get_param("divergent_cols"))(1000),
  focus_tfs_fn = get_param("scp_focus_tfs_fn"),
  plot_genes_2d = get_param("scp_plot_genes_2d"),
  meta_field_nm = get_param("scp_metadata_fig_factor"),
  highlight_cells_by_meta_field_nms = get_param("scp_highlight_cells_by_meta_field_nms"),
  mat_focus_genes_fn = get_param("scp_mat_focus_genes_fn"),
  show_2d_color_markers = get_param("scp_show_2d_color_markers"))
}
\arguments{
\item{outdir}{}

\item{mark_blacklist_terms}{}

\item{mark_colors_fn}{}

\item{fig_width_confu}{}

\item{fig_mat_per_clust_genes}{}

\item{fig_mat_gene_min_fold}{}

\item{fig_mat_gene_min_cov}{}

\item{fig_mat_width}{}

\item{fig_mat_height}{}

\item{fig_mat_text_cex}{}

\item{fig_mat_smooth_n}{}

\item{fig_2d_height}{}

\item{fig_2d_width}{}

\item{K_2dproj}{}

\item{K_cells_2dproj}{}

\item{force_max_deg_2dproj}{}

\item{T_edge_2dproj}{}

\item{T_edge_asym}{}

\item{restrict_edges_by_fpcor}{}

\item{expand_K_connect_2dproj}{}

\item{fold_shades}{}

\item{focus_tfs_fn}{}

\item{plot_genes_2d}{}

\item{meta_field_nm}{}

\item{highlight_cells_by_meta_field_nms}{}

\item{mat_focus_genes_fn}{}

\item{show_2d_color_markers}{}

\item{scdb.}{initialized scdb object}

\item{sc_cl.}{cell modules object to use for plotting. Ignored if plot object is available from scdb, search it in scdb if sc_cl = NULL}
}
\value{
returns invisibly a tgScMatClust2D plotting object.
}
\description{
sc_pipe_plots
}
