% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{sc_pipe_sort}
\alias{sc_pipe_sort}
\title{sc_pipe_sort}
\usage{
sc_pipe_sort(scdb, sc_mat = NULL, sort_types_markers_fn,
  sort_ref_centers_pref_fn,
  factor_by_median = get_param("scc_factor_by_median"),
  norm_size = get_param("scc_norm_size"),
  k_scale_umi = get_param("scc_k_scale_umi"), out_base_dir,
  out_base_dir_add_timestamp = F)
}
\arguments{
\item{sc_mat_tables_fn_pref}{}
}
\value{
returns invisibly a the original tgScMat with assigned types in @cell_metadata$assignd_type
}
\description{
Assign cells to types by finding the nearest center per cell
}
