% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorter.r
\name{sc_plot_sort_markers_on_clusts}
\alias{sc_plot_sort_markers_on_clusts}
\title{Plots markers x clusters, supporting addition of additional similiar markers.}
\usage{
sc_plot_sort_markers_on_clusts(sc_cl, markers_fn,
  remove_clustering_outliers = T, type = "clust_fp",
  min_max_clust_fp_to_report = get_param("scc_min_max_clust_fp_to_report"),
  add_genes_by_cor_above = NA, add_n_similar_genes = 0, k_scale_umi = 7,
  max_fp = 2, width = 2500, height = 2500, res = 220, cl_ord = NULL)
}
\arguments{
\item{sc_cl}{}

\item{markers_fn}{}

\item{type}{}

\item{add_genes_by_cor_above}{}

\item{add_n_similar_genes}{}

\item{k_scale_umi}{}

\item{max_fp}{}

\item{width}{}

\item{height}{}

\item{res}{}
}
\description{
Plots markers x clusters, supporting addition of additional similiar markers.
}
