% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorter.r
\name{sc_proj_mat_on_marks}
\alias{sc_proj_mat_on_marks}
\title{scrdb cell sorter}
\usage{
sc_proj_mat_on_marks(mat, markers_fn, factor_by_median = T, norm_size = T,
  k_scale_umi = 7)
}
\arguments{
\item{mat}{}

\item{k_scale_umi}{}
}
\description{
given a matrix and a bunch of marker sets, this function split cells by marker expression, and write separate matrices for downstream analysis
}
