% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{scc_bootstrap_clusts}
\alias{scc_bootstrap_clusts}
\title{bootstrap clustering: sample cells, cluster them (using the supplied scl object clustering parameters), build co-clust matrix and hclust that matrix}
\usage{
scc_bootstrap_clusts(scl, scdb = NULL,
  k_cut = get_param("scc_bootstrap_k_cut"),
  N_boot = get_param("scc_bootstrap_n_boot"),
  boot_ratio = get_param("scc_bootstrap_ratio"),
  n_procs = get_param("scc_bootstrap_n_procs"),
  min_clust_size = get_param("scc_bootstrap_min_clust_size"),
  verbose = get_param("scc_bootstrap_verbose"), rseed = 1)
}
\arguments{
\item{rseed}{}

\item{scl.}{Initial knn clustering result.}

\item{scdb.}{scdb instance, will save co-clust matrix there if supplied.}

\item{k_cut.}{Number of clusters to extract from the co-clust hclust tree}

\item{N_boot.}{Number of boostrap iterations}

\item{boot_ratio.}{Fraction of cells to sample in each iteration.}

\item{n_procs.}{Number of processes to use}

\item{min_clust_size.}{Pool cells in clusters smaller than min_clust_size to a single outlier cluster.}

\item{verbose.}{print progress messages.}
}
\description{
bootstrap clustering: sample cells, cluster them (using the supplied scl object clustering parameters), build co-clust matrix and hclust that matrix
}
