% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{scc_cluster_kmeans}
\alias{scc_cluster_kmeans}
\title{Cluster single cell data using kmeans}
\usage{
scc_cluster_kmeans(scmat, features = NULL, K = NA, norm_features = T,
  k_scale_umi = get_param("scc_k_scale_umi"),
  factor_by_median = get_param("scc_factor_by_median"),
  norm_size = get_param("scc_norm_size"), rseed = 1)
}
\arguments{
\item{scmat}{A tgScMat expression matrix.}

\item{features}{Either  a matrix of cell features (gene expression, PCA principle componant, etc.),
or a list of markers to be extracted from scmat. Clustering will be performed on these features.
If left blank, markers will be selected using the default method.}

\item{K}{number of clusters}

\item{norm_features}{Apply normalization on features}

\item{k_scale_umi}{}

\item{factor_by_median}{}

\item{norm_size}{}

\item{rseed}{}
}
\value{
A tgScMatClust object
}
\description{
Cluster single cell data using kmeans
}
