% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{scc_cluster_knn_graph}
\alias{scc_cluster_knn_graph}
\title{Cluster single cell data using KNN graph}
\usage{
scc_cluster_knn_graph(scmat, features = NULL, norm_features = T,
  k_scale_umi = get_param("scc_k_scale_umi"),
  min_clust_size = get_param("scc_min_clust_size"),
  use_balanced_knn = get_param("scc_use_balanced_knn"),
  k_knn = get_param("scc_k_knn"),
  ratio = get_param("scc_clust_size_to_k_knn_ratio"),
  do_consolidation = get_param("scc_knn_consolidate"), rseed = 1)
}
\arguments{
\item{scmat}{A tgScMat expression matrix.}

\item{features}{Either  a matrix of cell features (gene expression, PCA principle componant, etc.),
or a list of markers to be extracted from scmat. Clustering will be performed on these features.
If left blank, markers will be selected using the default method.}

\item{norm_features}{}

\item{k_scale_umi}{}

\item{min_clust_size}{}

\item{use_balanced_knn}{}

\item{k_knn}{}

\item{do_consolidation}{}

\item{rseed}{}
}
\value{
A tgScMatClust object
}
\description{
Cluster single cell data using KNN graph
}
