% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{scc_coclust_to_cell_modules}
\alias{scc_coclust_to_cell_modules}
\title{Update cell modules by a cut tree on a cells' co-clust matrix}
\usage{
scc_coclust_to_cell_modules(scdb, scl, cc, k_cut, min_clust_size,
  method = "hclust")
}
\arguments{
\item{scdb}{}

\item{scl}{}

\item{cc}{}

\item{k_cut}{}

\item{min_clust_size}{}
}
\description{
Update cell modules by a cut tree on a cells' co-clust matrix
}
