% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{scc_filter_bad_clusters}
\alias{scc_filter_bad_clusters}
\title{Filter bad clusters. Either clusters with out high enough fold-change genes, and/or clusters with low self-confusion neighbours}
\usage{
scc_filter_bad_clusters(scl,
  min_max_fold_change = get_param("scc_min_max_clust_fp_to_filter"),
  confu = NULL, k_knn_for_confu = get_param("scc_k_knn_for_confu"),
  min_self_confusion = get_param("scc_min_self_confusion"))
}
\arguments{
\item{scl}{}

\item{min_max_fold_change}{}

\item{confu}{}

\item{k_knn_for_confu}{}

\item{min_self_confusion}{}
}
\value{
tgScMatClust object with updated good clusters
}
\description{
Filter bad clusters. Either clusters with out high enough fold-change genes, and/or clusters with low self-confusion neighbours
}
