% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{scc_find_clustering_outliers}
\alias{scc_find_clustering_outliers}
\title{Find cells that are outliers in the current clusters}
\usage{
scc_find_clustering_outliers(scl,
  T_lfc = get_param("scc_clust_outliers_t_lfc"),
  min_outlier_u = get_param("scc_clust_outliers_min_gene_count"))
}
\arguments{
\item{scl}{}

\item{T_lfc}{}

\item{min_outlier_u}{}
}
\value{
vector of outlier cells
}
\description{
Find cells that are outliers in the current clusters
}
