% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{scc_import_clust}
\alias{scc_import_clust}
\title{Read a clustering result from file.}
\usage{
scc_import_clust(file, scmat = NULL)
}
\arguments{
\item{file}{Name of the file for inputting.}

\item{scmat}{an instance of the relevant \code{\linkS4class{tgScMat}}, optional.}
}
\description{
Read a clustering result from file.
}
