% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{scc_move_cells_to_orphan}
\alias{scc_move_cells_to_orphan}
\title{Remove bad cells to an 'orphan' cluster and update clusters slots}
\usage{
scc_move_cells_to_orphan(scl, bad_cells)
}
\arguments{
\item{scl}{}

\item{bad_cells}{}
}
\description{
Remove bad cells to an 'orphan' cluster and update clusters slots
}
