% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{scc_orphan_unlinked_cells}
\alias{scc_orphan_unlinked_cells}
\title{TODO: go over and remove/generelize specific params}
\usage{
scc_orphan_unlinked_cells(scl, k_knn = get_param("scc_k_knn"),
  min_genemod_fold = 1.5,
  min_confu_score = get_param("scc_min_confusion_score"),
  plot_dir = get_param("outdir"))
}
\arguments{
\item{scl}{}

\item{k_knn}{}

\item{min_genemod_fold}{}

\item{min_confu_score}{}

\item{plot_dir}{}
}
\value{

}
\description{
TODO: go over and remove/generelize specific params
}
