% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{scc_write_clust_fp_tab}
\alias{scc_write_clust_fp_tab}
\title{Output genes cell modules footprint matrix with metadata on the cell modules}
\usage{
scc_write_clust_fp_tab(sc_cl, outdir = get_param("outdir"),
  tab_clust_fp_fn = get_param("scc_clust_fp_fn"),
  min_max_clust_fp_to_report = get_param("scc_min_max_clust_fp_to_report"),
  clust_fp_metadata_fields = get_param("scc_clust_fp_metadata_fields"))
}
\arguments{
\item{sc_cl}{}

\item{tab_clust_fp_fn}{}

\item{min_max_clust_fp_to_report}{}

\item{clust_fp_metadata_fields}{}
}
\description{
Output genes cell modules footprint matrix with metadata on the cell modules
}
