% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_get_existing_obj_names}
\alias{scdb_get_existing_obj_names}
\title{Get names of objects saved under the scdb instance}
\usage{
scdb_get_existing_obj_names(scdb, pattern = NULL)
}
\arguments{
\item{scdb}{}

\item{pattern}{to filter output object names}
}
\value{
vector of object names
}
\description{
Get names of objects saved under the scdb instance
}
