% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_init}
\alias{scdb_init}
\title{Initialize tgScDbMngr object}
\usage{
scdb_init(basedir = get_param("scdb_basedir"),
  fn_prefix = get_param("scdb_fn_prefix"),
  which_not_lazy = get_param("scdb_which_not_lazy"))
}
\arguments{
\item{which_not_lazy}{}
}
\value{
invisible tgScDbMngr object
}
\description{
Initialize tgScDbMngr object
}
