% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_load}
\alias{scdb_load}
\title{Loads an object}
\usage{
scdb_load(scdb, obj_type)
}
\arguments{
\item{scdb}{}

\item{obj_type}{}
}
\value{
loaded object or NULL if it doesn't exist or marked not to lazy load
}
\description{
Loads an object
}
