% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_save}
\alias{scdb_save}
\title{Saves an object}
\usage{
scdb_save(scdb, object, obj_type, overwrite_if_exists = T)
}
\arguments{
\item{scdb}{}

\item{object}{}

\item{obj_type}{}

\item{overwrite_if_exists}{}
}
\description{
Saves an object
}
