% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{scdb_set_lazy}
\alias{scdb_set_lazy}
\title{Sets lazy load value for an object type}
\usage{
scdb_set_lazy(scdb, obj_type, lazy)
}
\arguments{
\item{scdb}{}

\item{obj_type}{}

\item{lazy}{T/F}
}
\description{
Sets lazy load value for an object type
}
