% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mods.r
\name{scgm_import}
\alias{scgm_import}
\title{Read a gmod object from file}
\usage{
scgm_import(file, scmat = NULL)
}
\arguments{
\item{file}{Name of the file for inputting.}

\item{scmat}{an instance of the relevant \code{\linkS4class{tgScMat}}, optional.}
}
\description{
Read a gmod object from file
}
