% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mods.r
\name{scgm_plot_gmod_on_2d}
\alias{scgm_plot_gmod_on_2d}
\title{Plot total gmod expression per cell on the 2d projection}
\usage{
scgm_plot_gmod_on_2d(sc_2d, gm, output_dir = sprintf("\%s/gmod_2d",
  get_param("outdir")), name = "", gm_labels = NULL, colspec = c("white",
  RColorBrewer::brewer.pal(9, "YlOrRd")), nshades = 1000, width = 800,
  height = 800, cex = 0.5, lwd = 0.5, zlim = NULL, zlim_q = 1e-04,
  show_top_n_genes = 30)
}
\arguments{
\item{sc_2d}{}

\item{gm}{}

\item{output_dir}{}

\item{name}{}

\item{gm_labels}{}

\item{colspec}{}

\item{nshades}{}

\item{width}{}

\item{height}{}

\item{cex}{}

\item{lwd}{}

\item{zlim}{}
}
\description{
Plot total gmod expression per cell on the 2d projection
}
