% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mods.r
\name{scgm_plot_matrices}
\alias{scgm_plot_matrices}
\title{Title}
\usage{
scgm_plot_matrices(gm, base_dir, name = "", w = 1000, h = 1000,
  h_fp = h, cex_gene = 0.6, filter_mods = NA, mark_mods = F,
  colspec = get_param("divergent_cols"), zlim = NULL, zlim_q = 1e-04,
  res = 180)
}
\arguments{
\item{gm}{}

\item{base_dir}{}

\item{w}{}

\item{h}{}

\item{h_fp}{}

\item{cex_gene}{}

\item{filter_mods}{}

\item{mark_mods}{}

\item{colspec}{}
}
\description{
Title
}
