% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_add_facs_indices_from_tables}
\alias{scm_add_facs_indices_from_tables}
\title{(optional) Load FACS indices info and add it to cell_metadata table}
\usage{
scm_add_facs_indices_from_tables(scmat,
  base_dir = get_param("scm_facs_idx_base_dir"),
  facs_idx_batches = get_param("scm_facs_idx_batches"),
  wells2cells_fn = get_param("scm_facs_idx_wells2cells_fn"),
  wells2cells_batch_field = get_param("scm_facs_idx_batch_field"),
  wells2cells_cell_field = get_param("scm_facs_idx_cell_field"),
  wells2cells_well_field = get_param("scm_facs_idx_well_field"),
  skip_missing_batch_facs_indices = get_param("scm_facs_idx_skip_missing"))
}
\arguments{
\item{scmat}{}

\item{base_dir}{}

\item{facs_idx_batches}{}

\item{wells2cells_fn}{}

\item{wells2cells_batch_field}{}

\item{wells2cells_cell_field}{}

\item{wells2cells_well_field}{}
}
\description{
(optional) Load FACS indices info and add it to cell_metadata table
}
