% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_import_mat}
\alias{scm_import_mat}
\title{Read a tgScMat object that saved to disk.}
\usage{
scm_import_mat(file)
}
\arguments{
\item{file}{Prefix of file names input (expecting %s.scmat and %s.scmat_cell_metadata and $s.scmat_attr to exist)}
}
\description{
Read a tgScMat object that saved to disk.
}
