% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_read_scmat_10x}
\alias{scm_read_scmat_10x}
\title{Read a matrix from the output of a 10x run. Extracting batch info from the cell barcodes (e.g. AACAGTGCGTGA-1 is in batch 1).}
\usage{
scm_read_scmat_10x(sparse_matrix_ifn = get_param("scm_10x_mat_ifn"),
  genes_ifn = get_param("scm_10x_mat_rownames_ifn"),
  cells_ifn = get_param("scm_10x_mat_colnames_ifn"),
  paralogs_policy = get_param("scm_10x_paralogs_policy"),
  use_batches = get_param("scm_10x_batches"),
  min_umis_n = get_param("scm_min_cell_umis"))
}
\arguments{
\item{sparse_matrix_ifn}{sparse matrix input file name. Expecting a tab delimited file with 3 columns and a header: 'row', 'column', 'value'. row and column fields are the row and column number of the value. Expecting 3 other files (if matrix file name is xxx.tab): xxx.rownames.tab, xxx.colnames.tab (each header by 'rowname' or 'colname'), with the row/column names. xxx.dims.tab (headed by 'dims') with 2 rows - the nmber of rows and the number of columns.}

\item{use_batches}{}

\item{min_umis_n}{minimum number of umi's for retaining the cell in the matrix (default: 200)}
}
\description{
Read a matrix from the output of a 10x run. Extracting batch info from the cell barcodes (e.g. AACAGTGCGTGA-1 is in batch 1).
}
