% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_remove_ambient_by_epsilon}
\alias{scm_remove_ambient_by_epsilon}
\title{Clean ambient noise from matrix}
\usage{
scm_remove_ambient_by_epsilon(scmat,
  epsilon = get_param("scm_amb_noise_epsilon"),
  metadata_batch = get_param("scm_batch_meta_attr"),
  min_umi_n = get_param("scm_min_cell_umi"))
}
\arguments{
\item{scmat}{A tgScMat object.}

\item{epsilon}{}

\item{metadata_batch}{}

\item{min_umi_n}{}
}
\description{
Clean ambient noise from matrix
}
