% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{scm_sub_mat}
\alias{scm_sub_mat}
\title{Extract sub-matrix}
\usage{
scm_sub_mat(scmat, genes = NULL, cells = NULL)
}
\arguments{
\item{scmat}{A tgScMat object.}

\item{genes}{Genes range.}

\item{cells}{Cells range}
}
\description{
Extract sub-matrix
}
