% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_layout.r
\name{scp_compute_cell_order}
\alias{scp_compute_cell_order}
\title{Compute cells order either by confusion matrix or by selected marker genes}
\usage{
scp_compute_cell_order(scm2d, K = -1, use_confu = F,
  per_clust_genes = get_param("scp_ord_per_clust_genes"),
  gene_min_fold = get_param("scp_ord_gene_min_fold"),
  gene_min_cov = get_param("scp_ord_gene_min_cov"),
  ignore_blist_genes = get_param("scp_ord_ignore_blist_genes"))
}
\arguments{
\item{scm2d}{}

\item{K}{}

\item{use_confu}{}

\item{per_clust_genes}{}

\item{gene_min_fold}{}

\item{gene_min_cov}{}

\item{ignore_blist_genes}{}
}
\description{
Compute cells order either by confusion matrix or by selected marker genes
}
