% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_layout.r
\name{scp_compute_cell_order_confu}
\alias{scp_compute_cell_order_confu}
\title{Order cells by confusion matrix}
\usage{
scp_compute_cell_order_confu(scm2d, K = -1)
}
\arguments{
\item{scm2d}{}

\item{K}{}
}
\description{
Order cells by confusion matrix
}
