% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_layout.r
\name{scp_compute_cell_order_markers}
\alias{scp_compute_cell_order_markers}
\title{Order cells by hclust on marker genes}
\usage{
scp_compute_cell_order_markers(scm2d, per_clust_genes = 5,
  gene_min_fold = 2.5, gene_min_cov = 0.5, ignore_genes = NULL)
}
\arguments{
\item{scm2d}{}

\item{per_clust_genes}{}

\item{gene_min_fold}{}

\item{gene_min_cov}{}

\item{ignore_genes}{}
}
\description{
Order cells by hclust on marker genes
}
\examples{

}
