% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_layout.r
\name{scp_init_plot}
\alias{scp_init_plot}
\title{Init a single cell plotting object}
\usage{
scp_init_plot(scc)
}
\arguments{
\item{scc}{A tgScMatClust object, created using scc_cluster().}
}
\value{
A tgScMatClustLayout2D object, used in all of the plotting functions.
}
\description{
Init a single cell plotting object, tgScMatClust2D
}
