% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_layout.r
\name{scp_plot_confusion_mat}
\alias{scp_plot_confusion_mat}
\title{Plotting a the cluster confusion matrix}
\usage{
scp_plot_confusion_mat(scm2d, outdir = ".", K = -1, width = 1000,
  height = 1000)
}
\arguments{
\item{scm2d}{an object of clust layout}

\item{outdir}{- file location}

\item{K}{(20) - the parameter}
}
\description{
Plotting a the cluster confusion matrix
}
