% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_layout.r
\name{scp_plot_mat}
\alias{scp_plot_mat}
\title{Plotting a heatmap of clusters}
\usage{
scp_plot_mat(scl2d, png_fn = NULL, plot_cells = T, genes_pool = NULL,
  blacklist_genes = c(), remove_blacklist = T, width = 2000,
  height = NA, text_cex = 1, alt_side_text = F, per_clust_genes = 5,
  gene_min_fold = 2.5, gene_min_cov = 0.5, norm_type = "linear",
  tot_umi_scale = 1000, smooth_n = 5, top_marg = c(0, 13, 5, 20),
  lower_marg = c(5, 13, 0, 20), fp_shades = colorRampPalette(c("white",
  "lightsteelblue", "blue", "darkblue", "purple", "brown", "black",
  "yellow"))(1000), cell_batch = NULL, batch_color = NULL,
  cell_ord = NULL, genes = NULL)
}
\arguments{
\item{png_fn}{}

\item{plot_cells}{}

\item{genes_pool}{}

\item{blacklist_genes}{}

\item{width}{}

\item{height}{}

\item{text_cex}{}

\item{alt_side_text}{}

\item{per_clust_genes}{}

\item{gene_min_fold}{}

\item{gene_min_cov}{}

\item{norm_type}{}

\item{tot_umi_scale}{}

\item{smooth_n}{}

\item{top_marg}{}

\item{lower_marg}{}

\item{fp_shades}{}

\item{cell_batch}{}

\item{batch_color}{}

\item{cell_ord}{}

\item{genes}{}
}
\description{
Plotting a heatmap of clusters
}
