% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_select.r
\name{select_markers}
\alias{select_markers}
\title{Gene selection}
\usage{
select_markers(scmat, gstat = NULL, type = c("any", "niche", "varmean",
  "sizecor"), ...)
}
\arguments{
\item{scmat}{tgScMat data matrix}

\item{type}{.}

\item{...}{.}
}
\description{
Strategies for gene selection given a single cell matrix
}
