% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\docType{methods}
\name{show,tgScDbMngr-method}
\alias{show,tgScDbMngr-method}
\title{Print scdb object}
\usage{
\S4method{show}{tgScDbMngr}(object)
}
\arguments{
\item{tgScGeneMods}{}
}
\description{
Print scdb object
}
