% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_clust.r
\name{tg_piecewise_graph_cover}
\alias{tg_piecewise_graph_cover}
\title{Generic graph cluster-cover}
\usage{
tg_piecewise_graph_cover(m_e, min_clust_size, verbose = F, consolidate = T)
}
\arguments{
\item{m_e}{the adjacency matrix of the graph to be clustered}

\item{min_clust_size}{used to filter small clusters at the end of the process}
}
\description{
Generic graph cluster-cover
}
