% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_clust.r
\name{tg_simmat_to_knn_ordered}
\alias{tg_simmat_to_knn_ordered}
\title{generate a k_nn graph from a similarity matrix}
\usage{
tg_simmat_to_knn_ordered(mat, k_knn)
}
\arguments{
\item{m_knn}{}
}
\description{
generate a k_nn graph from a similarity matrix
}
