% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\docType{class}
\name{tgScDbMngr-class}
\alias{tgScDbMngr-class}
\alias{tgScDbMngr}
\title{Loading and saving scrdb objects}
\description{
Loading and saving scrdb objects
}
\section{Slots}{

\describe{
\item{\code{basedir}}{character. Directory containing the objects for save and load.}

\item{\code{fn_prefix}}{character. Filename prefix for the objects managed by the instance.}

\item{\code{which_not_lazy}}{vector. Object types not to lazy load (regenerate object even if it can be loaded from the db)}
}}

