% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mods.r
\docType{class}
\name{tgScGeneMods-class}
\alias{tgScGeneMods-class}
\alias{tgScGeneMods}
\title{Gene modules (groups of correlated genes)
Supporting algorithms for detection of modules and some stats on them}
\description{
Gene modules (groups of correlated genes)
Supporting algorithms for detection of modules and some stats on them
}
\section{Slots}{

\describe{
\item{\code{scmat}}{tgScMat.}

\item{\code{foc_genes}}{vector. selected genes}

\item{\code{feat_mat}}{matrix.}

\item{\code{gene_cor}}{matrix.}

\item{\code{rnd_gene_cor}}{matrix.}

\item{\code{hc_order}}{vector.}

\item{\code{gmods}}{vector.}

\item{\code{indep_genes}}{vector.}

\item{\code{gmod_mean}}{matrix.}

\item{\code{gmod_fp}}{matrix.}

\item{\code{alg_params}}{list.}

\item{\code{nclust}}{integer.}

\item{\code{knn_ordered}}{matrix. k_knn X ncells matrix,  column i contains cell i's k nearest neighbors, ordered from closest to farthest.}

\item{\code{m_knn}}{matrix.}

\item{\code{rseed}}{numeric.}
}}

