% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mods.r
\docType{methods}
\name{tgscm_export,tgScGeneMods-method}
\alias{tgscm_export,tgScGeneMods-method}
\title{Export a gene mod object to files}
\usage{
\S4method{tgscm_export}{tgScGeneMods}(.Object, file, supress_mat = F, ...)
}
\arguments{
\item{file}{Prefix of file names for outputting.}
}
\description{
Export a gene mod object to files
}
