% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\docType{methods}
\name{tgscm_export,tgScMatClust-method}
\alias{tgscm_export,tgScMatClust-method}
\title{Export a clustering results to file.}
\usage{
\S4method{tgscm_export}{tgScMatClust}(.Object, file, supress_mat = F, ...)
}
\arguments{
\item{file}{Prefix of file names for outputting.}
}
\description{
Export a clustering results to file.
}
