% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmat.r
\name{tgscm_export}
\alias{tgscm_export}
\title{Export a matrix to file.}
\usage{
tgscm_export(.Object, file, ...)
}
\arguments{
\item{file}{Prefix of file names for outputting.}
}
\description{
Export a matrix to file.
}
