% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.r
\name{tgscm_reduce}
\alias{tgscm_reduce}
\title{Dimensionality reduction for gene expression}
\usage{
tgscm_reduce(.Object, type, d_n, ...)
}
\arguments{
\item{type}{.}

\item{d_n}{.}

\item{Object}{.}
}
\description{
Reduce a single cell matrix over a gene set, or through various dimensionality reduction strategies.
}
