% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\docType{class}
\name{tgScMatClust-class}
\alias{tgScMatClust-class}
\alias{tgScMatClust}
\title{A clustered matrix
Representing a clustered matrix, and implementing several basic algorithms for constructing it
from a matrix}
\description{
A clustered matrix
Representing a clustered matrix, and implementing several basic algorithms for constructing it
from a matrix
}
\section{Slots}{

\describe{
\item{\code{scmat}}{tgScMat.}

\item{\code{feat_mat}}{Matrix.}

\item{\code{clusts}}{vector.}

\item{\code{orphan_clust_id}}{numeric.}

\item{\code{clust_fp}}{matrix.}

\item{\code{clust_gcov}}{matrix.}

\item{\code{alg_params}}{list.}

\item{\code{nclust}}{integer.}

\item{\code{knn_ordered}}{matrix. k_knn X ncells matrix,  column i contains cell i's k nearest neighbors, ordered fronm closest to farthest.}

\item{\code{knn_is_balanced}}{logical.}

\item{\code{m_knn}}{dgCMatrix.}

\item{\code{rseed}}{numeric.}
}}

