% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_layout.r
\docType{class}
\name{tgScMatClustLayout2D-class}
\alias{tgScMatClustLayout2D-class}
\alias{tgScMatClustLayout2D}
\title{Plotting object for a clustering solution}
\description{
Plotting object for a clustering solution

Representing a 2d layout for a cluster model, used for plotting.

\code{tgScMatClustLayout2D} represents a 2D layout for a cluster model.
}
\section{Slots}{

\describe{
\item{\code{scl}}{tgScMatClust.}

\item{\code{cl_ord}}{vector.}

\item{\code{c_ord}}{vector.}

\item{\code{cl_colors}}{vector.}

\item{\code{c_colors}}{vector.}

\item{\code{mark_colors}}{vector.}

\item{\code{mark_groups}}{vector}

\item{\code{current_subset}}{vector.}

\item{\code{x_cl}}{vector.}

\item{\code{y_cl}}{vector.}

\item{\code{clust_graph}}{matrix.}

\item{\code{x}}{vector.}

\item{\code{y}}{vector.}

\item{\code{cell_2d_sd}}{vector.}

\item{\code{blist_genes}}{vector.}
}}

