% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Sample raw UMI counts}
\format{An object of class \code{matrix} with 5782 rows and 200 columns.}
\usage{
toydata
}
\description{
A matrix containing sample UMI counts
}
\keyword{datasets}
